beagle_x15: switch to /usr/bin/bb_dnsmasq_config.sh

+++ boot/beagle_x15.sh
@@ -380,25 +380,29 @@ if [ "x${dnsmasq_usb0_usb1}" = "xenable" ] ; then
 			echo "${log} dnsmasq: setting up for usb0/usb1"
 			disable_connman_dnsproxy
 
-			wfile="/etc/dnsmasq.d/SoftAp0"
-			echo "interface=usb0" > ${wfile}
-			echo "interface=usb1" >> ${wfile}
-			echo "port=53" >> ${wfile}
-			echo "dhcp-authoritative" >> ${wfile}
-			echo "domain-needed" >> ${wfile}
-			echo "bogus-priv" >> ${wfile}
-			echo "expand-hosts" >> ${wfile}
-			echo "cache-size=2048" >> ${wfile}
-			echo "dhcp-range=usb0,192.168.7.1,192.168.7.1,2m" >> ${wfile}
-			echo "dhcp-range=usb1,192.168.6.1,192.168.6.1,2m" >> ${wfile}
-			echo "listen-address=127.0.0.1" >> ${wfile}
-			echo "listen-address=192.168.7.2" >> ${wfile}
-			echo "listen-address=192.168.6.2" >> ${wfile}
-			echo "dhcp-option=usb0,3" >> ${wfile}
-			echo "dhcp-option=usb0,6" >> ${wfile}
-			echo "dhcp-option=usb1,3" >> ${wfile}
-			echo "dhcp-option=usb1,6" >> ${wfile}
-			echo "dhcp-leasefile=/var/run/dnsmasq.leases" >> ${wfile}
+			if [ -f /usr/bin/bb_dnsmasq_config.sh ] ; then
+				/usr/bin/bb_dnsmasq_config.sh || true
+			else
+				wfile="/etc/dnsmasq.d/SoftAp0"
+				echo "interface=usb0" > ${wfile}
+				echo "interface=usb1" >> ${wfile}
+				echo "port=53" >> ${wfile}
+				echo "dhcp-authoritative" >> ${wfile}
+				echo "domain-needed" >> ${wfile}
+				echo "bogus-priv" >> ${wfile}
+				echo "expand-hosts" >> ${wfile}
+				echo "cache-size=2048" >> ${wfile}
+				echo "dhcp-range=usb0,192.168.7.1,192.168.7.1,2m" >> ${wfile}
+				echo "dhcp-range=usb1,192.168.6.1,192.168.6.1,2m" >> ${wfile}
+				echo "listen-address=127.0.0.1" >> ${wfile}
+				echo "listen-address=192.168.7.2" >> ${wfile}
+				echo "listen-address=192.168.6.2" >> ${wfile}
+				echo "dhcp-option=usb0,3" >> ${wfile}
+				echo "dhcp-option=usb0,6" >> ${wfile}
+				echo "dhcp-option=usb1,3" >> ${wfile}
+				echo "dhcp-option=usb1,6" >> ${wfile}
+				echo "dhcp-leasefile=/var/run/dnsmasq.leases" >> ${wfile}
+			fi
 
 			systemctl restart dnsmasq || true
 		else
